+++ sdk/servicebus/azure-servicebus/tests/test_sb_client.py
@@ -14,7 +14,7 @@ from datetime import datetime, timedelta
 from azure.common import AzureHttpError, AzureConflictHttpError
 from azure.mgmt.servicebus.models import AccessRights
 from azure.servicebus import ServiceBusClient, ServiceBusSender
-from azure.servicebus._base_handler import ServiceBusSharedKeyCredential
+from azure.servicebus import ServiceBusSharedKeyCredential
 from azure.servicebus._common.message import Message, PeekedMessage
 from azure.servicebus.exceptions import (
     ServiceBusError,